server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "opencv-python>=4.8.0",
    "pillow>=9.5.0,<11.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.32.5",
    "simple-lama-inpainting>=0.1.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.3.248",
    "uvicorn>=0.32.0",
    "websockets>=14.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
